Fixed an error where an non authenticated user would cause an error on the reviews

+++ resources/views/partials/review.blade.php
@@ -18,11 +18,13 @@
             <div class="col text-center">
                 {{ $review->getDate() }}
             </div>
-            @if (Auth::user()["user_id"] ==  $review->user()[0]["user_id"])
-                <div class="col-lg-1 d-flex pt-0">
-                    <button class="btn delete_review" style="background-color: transparent; color:red;" onclick={{"deleteReviewRequest(".$review["review_id"].")"}}><i class="bi bi-trash-fill"></i></button>
-                    <button class="btn edit_review" style="background-color: transparent; color:grey;" onclick={{"editReviewRequest(".$review["review_id"].")"}}><i class="bi bi-pen-fill"></i></button>
-                </div>
+            @if (Auth::check())
+                @if (Auth::user()["user_id"] ==  $review->user()[0]["user_id"])
+                    <div class="col-lg-1 d-flex pt-0">
+                        <button class="btn delete_review" style="background-color: transparent; color:red;" onclick={{"deleteReviewRequest(".$review["review_id"].")"}}><i class="bi bi-trash-fill"></i></button>
+                        <button class="btn edit_review" style="background-color: transparent; color:grey;" onclick={{"editReviewRequest(".$review["review_id"].")"}}><i class="bi bi-pen-fill"></i></button>
+                    </div>
+                @endif
             @endif
         </div>
     </div>
